requirements.txt
absl-py==1.4.0
appdirs==1.4.4
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
filelock==3.12.2
fsspec==2023.6.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.15.1
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.6.0
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.1rc1
Markdown==3.4.3
MarkupSafe==2.1.3
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.3
pooch==1.6.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
soundfile==0.12.1
soxr==0.3.5
speechbrain==0.5.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.1.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
